+++ NoodleExtensions/NoodleExtensions.csproj
@@ -71,6 +71,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Animation\AnimationController.cs" />
+    <Compile Include="Animation\AnimationHelper.cs" />
     <Compile Include="Animation\Easings.cs" />
     <Compile Include="Animation\Events\AnimateVisualTrack.cs" />
     <Compile Include="Animation\Events\AssignAnimation.cs" />
